Rename syntax file to exec
