+++ lib/exec.d.ts
