+++ lib/hc.js
@@ -1,12 +1,12 @@
 #!/usr/bin/env node
 "use strict";
-var syntax_1 = require("./syntax");
+var exec_1 = require("./exec");
 var input = "“Hello, MAML!”";
 var argv = process.argv;
 if (argv.length > 3) {
     input = argv[3];
 }
 ;
-var output = syntax_1.exec(input);
+var output = exec_1.exec(input);
 console.log(output);
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiaGMuanMiLCJzb3VyY2VSb290IjoiIiwic291cmNlcyI6WyIuLi9zcmMvaGMudHMiXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6Ijs7QUFFQSxtQ0FBZ0M7QUFFaEMsSUFBSSxLQUFLLEdBQUcsZ0JBQWdCLENBQUM7QUFDN0IsSUFBSSxJQUFJLEdBQUcsT0FBTyxDQUFDLElBQUksQ0FBQztBQUN4QixFQUFFLENBQUMsQ0FBQyxJQUFJLENBQUMsTUFBTSxHQUFHLENBQUMsQ0FBQyxDQUFDLENBQUM7SUFBQyxLQUFLLEdBQUcsSUFBSSxDQUFDLENBQUMsQ0FBQyxDQUFDO0FBQUMsQ0FBQztBQUFBLENBQUM7QUFFMUMsSUFBTSxNQUFNLEdBQUcsYUFBSSxDQUFDLEtBQUssQ0FBQyxDQUFDO0FBQzNCLE9BQU8sQ0FBQyxHQUFHLENBQUMsTUFBTSxDQUFDLENBQUMifQ==
+//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiaGMuanMiLCJzb3VyY2VSb290IjoiIiwic291cmNlcyI6WyIuLi9zcmMvaGMudHMiXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6Ijs7QUFFQSwrQkFBOEI7QUFFOUIsSUFBSSxLQUFLLEdBQUcsZ0JBQWdCLENBQUM7QUFDN0IsSUFBSSxJQUFJLEdBQUcsT0FBTyxDQUFDLElBQUksQ0FBQztBQUN4QixFQUFFLENBQUMsQ0FBQyxJQUFJLENBQUMsTUFBTSxHQUFHLENBQUMsQ0FBQyxDQUFDLENBQUM7SUFBQyxLQUFLLEdBQUcsSUFBSSxDQUFDLENBQUMsQ0FBQyxDQUFDO0FBQUMsQ0FBQztBQUFBLENBQUM7QUFFMUMsSUFBTSxNQUFNLEdBQUcsV0FBSSxDQUFDLEtBQUssQ0FBQyxDQUFDO0FBQzNCLE9BQU8sQ0FBQyxHQUFHLENBQUMsTUFBTSxDQUFDLENBQUMifQ==
